feat:댓글작성/삭제, 찜하기, 유저프로필페이지

+++ spartamarket/products/views.py
@@ -3,9 +3,13 @@ from django.shortcuts import(
     redirect,
     get_object_or_404,
     )
-from django.views.decorators.http import require_POST
-from .models import Product
-from .forms import ProductForm
+from django.views.decorators.http import (
+    require_POST,
+    require_http_methods
+    )
+from django.contrib.auth.decorators import login_required
+from .models import Product, Comment
+from .forms import ProductForm, CommentForm
 
 
 def index(request):
@@ -16,11 +20,14 @@ def index(request):
     return render(request, 'products/index.html', context)
 
 
+@login_required
 def create(request):
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
-            product = form.save()
+            product = form.save(commit=False)
+            product.author = request.user
+            product.save()
             return redirect('products:details', product.pk)
     else:
         form = ProductForm()
@@ -32,8 +39,12 @@ def create(request):
 
 def details(request, pk):
     product = get_object_or_404(Product, pk=pk)
+    comments = product.comments.all()
+    comment_form = CommentForm()
     context = {
         'product':product,
+        'comments':comments,
+        'comment_form':comment_form,
     }
     return render(request, 'products/details.html', context)
 
@@ -41,27 +52,63 @@ def details(request, pk):
 @require_POST
 def delete(request, pk):
     product = get_object_or_404(Product, pk=pk)
-    product.delete()
-    return redirect('index')
+    if request.user == product.author:
+        product.delete()
+        return redirect('index')
+    else:
+        return redirect('articles:details', product.pk)
 
 
+@require_http_methods(['GET', 'POST'])
 def update(request, pk):
     product = get_object_or_404(Product, pk=pk)
-    if request.method == 'POST':
-        form = ProductForm(request.POST, request.FILES, instance=product)
-        if form.is_valid():
-                product = form.save()
-                return redirect("products:details", product.pk)
+    if request.user == product.author:
+        if request.method == 'POST':
+            form = ProductForm(request.POST, request.FILES, instance=product)
+            if form.is_valid():
+                    product = form.save()
+                    return redirect("products:details", product.pk)
+        else:
+            form = ProductForm(instance=product)
+            
+        context = {
+            'product':product,
+            'form':form,
+        }
+        return render(request, 'products/update.html', context)
     else:
-        form = ProductForm(instance=product)
-    context = {
-        'product':product,
-        'form':form,
-    }
-    return render(request, 'products/update.html', context)
+        return redirect('articles:details', product.pk)
 
 
+@login_required
 def comments(request, pk):
     product = get_object_or_404(Product, pk=pk)
-    pass
+    form = CommentForm(request.POST)
+    if form.is_valid():
+        comment = form.save(commit=False)
+        comment.product = product
+        comment.author = request.user
+        comment.save()
+        return redirect("products:details", product.pk)
 
+
+@require_POST
+def comments_delete(request, pk):
+    comment = get_object_or_404(Comment, pk=pk)
+    product_pk = comment.product.pk
+    if request.user == comment.author:
+        comment.delete()
+    return redirect("products:details", pk=product_pk)
+
+
+@require_POST
+def like(request, pk):
+    if request.user.is_authenticated:
+        product = get_object_or_404(Product, pk=pk)
+        if product.like_users.filter(pk=request.user.pk).exists():
+            product.like_users.remove(request.user)
+        else:
+            product.like_users.add(request.user)
+            product.save()
+        return redirect('products:details', pk=pk)
+    return redirect('accounts:login')
